requirements.txt
asgiref==3.7.2; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
django==5.0; python_version >= '3.10'
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-yasg[validation]==1.21.7; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.20.0; python_version >= '3.8'
jsonschema-specifications==2023.11.2; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pillow==10.1.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.32.0; python_version >= '3.8'
rpds-py==0.13.2; python_version >= '3.8'
setuptools==69.0.2; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
swagger-spec-validator==3.0.3; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2023.3; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
whitenoise==6.6.0; python_version >= '3.8'

djangorestframework~=3.14.0
Django~=5.0